Removed viewmodel from the parameters of the 4 main screens
Now the viewmodels of the 4 main screen (the screen shown by the bottom navigation bar) are not passed as parameters of the screen. Just the state and the callback are passed to the screen. This will simplify the unit test of the screen since can be injected a dummy state